reorganized portfolio and user components and reimported correctly
