+++ client/src/components/UserComponents/UserPortfolioValue/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-function PortfolioValue({value}) {
+function UserPortfolioValue({value}) {
     return(
         <p>
             $12312312
@@ -8,4 +8,4 @@ function PortfolioValue({value}) {
     );
 }
 
-export default PortfolioValue;
+export default UserPortfolioValue;
