+++ client/src/pages/Portfolio.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import PortfolioColumns from "../components/Portfoliocolumns/Portfoliocolumns";
+import PortfolioColumns from "../components/PortfolioComponents/PortfolioColumns";
 import API from "./../utils/API";
 import { useAuth } from "../utils/auth";
 
@@ -8,7 +8,7 @@ function Portfolio() {
   const [email, setEmail] = useState("");
   const { user } = useAuth();
 
-  
+
 
   useEffect(() => {
     API.getUser(user.id).then((res) => {
